requirements.txt
# AWS CDK
aws-cdk-lib>=2.0.0
constructs>=10.0.0

# AWS SDK
boto3>=1.26.0
botocore>=1.29.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
torch>=2.0.0
transformers>=4.30.0
openai>=1.0.0

# Web
streamlit>=1.25.0
plotly>=5.15.0

# Development
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Infrastructure
python-dotenv>=1.0.0 